QUESTS / Building better quest steps
I didn't like how I need to create an individual Prefab for EACH and EVERY quest step so I've change it so that all the steps get instiantiated when the Quest is first created, and uses information from the ScriptableObject to define those steps. Still needs testing but it should be ok. Need to destroy the steps when no longer needed aswell.