⚠️Swaps from SERVER authoritative to CLIENT authoritiative
This is a big decision because it will effect how the entire network code is written. Since this is a COOP game I feel that having things be Client auth is safe to do. (And any possible future competitive aspects will be among friends so that should be fine, right?)