🔀 I LOVE CREATING FUTURE MERGE CONFLICTS YIPEE
IMPORTANT: I have realized that in order to continue work on the multiplayer, I need to re-asses and re-design the way the map is presented and stored (256^2 gameObjects is very bad!!!) This will hopefully not be that long of a detour, and improve performance in the future. Doing this now WILL cause a merge conflict in the future though when I return to the MP stuff, but I haven't gone that deep into it yet so it shouldn't be too bad.