Start work on creating/visualizing the tilemap
Still buggy, gonna be annoying to do. Also important note, gonna be shifting ACCESS & DOOR logic to TileBlocks because its better to have it all in one place. I don't think it would be wise to do this to machines? Because despite their prevalence on the map, they are VERY complicated. We shall see.